Refactor warlock rotations to share base class
