+++ BotProfiles/WarlockDestruction/WarlockDestruction.csproj
@@ -9,6 +9,7 @@
   <ItemGroup>
     <ProjectReference Include="..\BotCommLayer\BotCommLayer.csproj" />
     <ProjectReference Include="..\BotRunner\BotRunner.csproj" />
+    <Compile Include="..\Common\WarlockBaseRotationTask.cs" />
   </ItemGroup>
 
   <ItemGroup>
